backend_native/requirements.txt
fastapi
uvicorn
redis
python-multipart
imagehash
Pillow
ultralytics
opencv-python
numpy
deepface
tf-keras
tensorflow-cpu
scipy
scikit-learn
